+++ packages/orq-rc/src/orq_ai_sdk/models/createidentityop.py
@@ -137,7 +137,7 @@ class CreateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-28T17:03:47.548Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-29T06:48:10.073Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")