+++ packages/orq-rc/RELEASES.md
@@ -5089,3 +5089,13 @@ Based on:
 - [python v4.3.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.7 - packages/orq-rc
+
+## 2026-01-29 06:51:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.700.2 (2.801.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.8] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.8] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.8 - packages/orq-rc